chore(ci): make helper integration tests runnable (#6562)
* chore(ci): make helper integration tests runnable

* boolean

* hop

* load integration env vars in test jobs

* v4

* v4

* like this?

---------

+++ packages/algoliasearch-helper/test/integration-utils.js
@@ -11,12 +11,50 @@ function setup(indexName, fn) {
     // all indexing requests must be done in https
     protocol: 'https:',
   });
-  client.deleteIndex =
-    client.deleteIndex ||
-    function (deleteIndexName) {
+  var hasInitIndex = Boolean(client.initIndex);
+  var originalDeleteIndex = client.deleteIndex
+    ? client.deleteIndex.bind(client)
+    : undefined;
+
+  client.deleteIndex = function (deleteIndexName) {
+    if (!originalDeleteIndex) {
       return client.initIndex(deleteIndexName).delete();
-    };
+    }
+    if (!hasInitIndex) {
+      return originalDeleteIndex({ indexName: deleteIndexName });
+    }
+    return originalDeleteIndex(deleteIndexName);
+  };
   client.listIndexes = client.listIndexes || client.listIndices;
+  client.initIndex =
+    client.initIndex ||
+    function (initIndexName) {
+      return {
+        addObjects: function (objects) {
+          return client.saveObjects({
+            objects: objects,
+            indexName: initIndexName,
+          });
+        },
+        clearObjects: function () {
+          return client.clearObjects({ indexName: initIndexName });
+        },
+        setSettings: function (settings) {
+          return client.setSettings({
+            indexName: initIndexName,
+            indexSettings: settings,
+          });
+        },
+        waitTask: function (taskID) {
+          return client.waitForTask({
+            indexName,
+            initIndexName,
+            taskID: taskID,
+          });
+        },
+      };
+    };
+  client.destroy = client.destroy || function () {};
 
   var index = client.initIndex(indexName);
   index.addObjects =
@@ -31,6 +69,9 @@ function setup(indexName, fn) {
         });
     };
   index.clearIndex = index.clearIndex || index.clearObjects;
+
+  // Ensure we go into the right integration branches (and client stays v5)
+  client.initIndex = hasInitIndex ? client.initIndex : undefined;
 
   return index
     .clearIndex()
